Ca avance pas mal (rendu dans le main, quelques tests qui ne fonctionnent pas encore)

+++ TP1-H20-Windows/TP1-H20/include/GestionnaireSaisons.h
@@ -7,7 +7,7 @@
 class GestionnaireSaisons
 {
 public:
-    static constexpr size_t SAISON_INEXSISTANTE = -1;
+    static constexpr int SAISON_INEXSISTANTE = -1;
 
     GestionnaireSaisons() = default;
     ~GestionnaireSaisons();
@@ -24,7 +24,7 @@ public:
     size_t getNbSaisons() const;
 
 protected:
-    size_t trouverIndexSaison(const unsigned int numSaison) const;
+    int trouverIndexSaison(const unsigned int numSaison) const;
 
     // Attribute
     std::vector<std::unique_ptr<Saison>> saisons_;
